Test tweaks for package split.
