+++ NOTICE
@@ -1,4 +1,4 @@
-php-wasm aka PIB
+php-wasm
 
 Copyright 2018-2020 Oraoto
 Copyright 2019-2020 Sean Morris
