+++ packages/freetype/test/basic.mjs
@@ -1,6 +1,6 @@
 import { test } from 'node:test';
 import { strict as assert } from 'node:assert';
-import { PhpNode } from '../../../PhpNode.mjs';
+import { PhpNode } from '../../packages/php-wasm/PhpNode.mjs';
 
 test('Function "imagettftext" exists.', async () => {
 	const php = new PhpNode();
